chore: refactor to introduce UserSettings page
