+++ src/pages/Settings/UserSettings.tsx
@@ -1,0 +1,7 @@
+import { SettingsPage } from './SettingsPage'
+
+export const UserSettings = (_props: { adminEditableUserId?: string }) => (
+  <>
+    <SettingsPage adminEditableUserId={_props.adminEditableUserId} />
+  </>
+)
