+++ src/pages/Settings/index.tsx
@@ -4,7 +4,8 @@ import { Route, Switch, withRouter } from 'react-router-dom'
 import { inject, observer } from 'mobx-react'
 import type { UserStore } from 'src/stores/User/user.store'
 import type { IUser } from 'src/models/user.models'
-import { UserSettings } from './SettingsPage'
+// import { UserSettings } from './SettingsPage'
+import { UserSettings } from './UserSettings'
 import { Text, Flex } from 'theme-ui'
 
 interface InjectedProps extends IProps {
